% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_alpha}
\alias{is_alpha}
\alias{is_digit}
\alias{is_alpha_digit}
\alias{is_whitespace}
\title{Predicate functions}
\usage{
is_alpha(x)

is_digit(x)

is_alpha_digit(x)

is_whitespace(x)
}
\arguments{
\item{x}{character}
}
\value{
logical
}
\description{
Predicate functions
}
